Refactor/#54: 커뮤세부 컴포넌트 파일 묶기
