+++ src/pages/CommunityDetail/components/comment/commentBoard/CommentBoard.styled.ts
