+++ src/pages/CommunityDetail/components/comment/commentCard/CommentCard.styled.ts
