+++ src/pages/CommunityDetail/components/input/commentInput/CommentInput.styled.ts
